requirements.txt
Package          Version
---------------- -------
aniso8601        9.0.1
autopep8         1.5.7
click            8.0.1
colorama         0.4.4
Flask            2.0.1
Flask-RESTful    0.3.9
Flask-SQLAlchemy 2.5.1
greenlet         1.1.1
itsdangerous     2.0.1
Jinja2           3.0.1
MarkupSafe       2.0.1
paho-mqtt        1.5.1
pip              21.2.4
pycodestyle      2.7.0
pytz             2021.1
setuptools       58.0.4
six              1.16.0
SQLAlchemy       1.4.23
toml             0.10.2
Werkzeug         2.0.1
wheel            0.37.0
